Simplify hosted EMG pose models
