+++ braindecode/modules/__init__.py
@@ -2,6 +2,7 @@ from .activation import (
     GatedLinearUnit,
     LogActivation,
     SafeLog,
+    SmoothMaximumUnit,
     Square,
 )
 from .attention import (
@@ -65,6 +66,7 @@ __all__ = [
     "GatedLinearUnit",
     "LogActivation",
     "SafeLog",
+    "SmoothMaximumUnit",
     "Square",
     "CAT",
     "CBAM",
